+++ admin_app/urls.py
@@ -3,5 +3,5 @@ from django.urls import path
 from admin_app.views import Franchise_TypeView
 
 urlpatterns = [
-    path("",Franchise_TypeView.as_view())
+    path("ftype/",Franchise_TypeView.as_view())
 ]
