+++ admin_app/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from .models import Franchise_Type
+from Accounts.models import Franchise_Type
 from rest_framework import serializers
 from .serializer import Franchise_Type_Serializer
 from rest_framework import status
